move to correct place
