+++ Content.Client/Administration/UI/EntitySearch/EntitySearchWindow.xaml
