+++ Content.Server/Administration/Commands/EntitySearchUiCommand.cs
@@ -1,4 +1,4 @@
-using Content.Server.Administration.UI;
+using Content.Server.Administration.UI.EntitySearch;
 using Content.Server.EUI;
 using Content.Shared.Administration;
 using Robust.Shared.Console;
